Support ui resources in src/main/java.
Configured resource-handling in maven build so that we now pick-up UI resources from src/main/java.
Previously we required users to build parallel package hierarchies in src/main/resources in order to support UI resources. This can be a bit cumbersome to work with when packages are refactored etc.

This fixes #4.